termination: refactor
