+++ generational_evolution_engine_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/aurelien-rainone/evolve/pkg/operator/mutation"
 	"github.com/aurelien-rainone/evolve/pkg/operator/xover"
 	"github.com/aurelien-rainone/evolve/pkg/selection"
-	"github.com/aurelien-rainone/evolve/termination"
+	"github.com/aurelien-rainone/evolve/pkg/termination"
 	"github.com/stretchr/testify/assert"
 )
