+++ pkg/termination/elapsed_time.go
