+++ pkg/termination/generation_count.go
