+++ pkg/termination/user_abort.go
